assert.NoError() -> require.NoError() in kad tests
For reasons unknowable, the authors of the testify lib chose to use the
word "assert" for specifying expectations, and the word "require" for
specifying assertions. This causes much trouble when people expect that
a failed AssertFoo() call will abort the test, and it doesn't.

It's probably too much to ask to get Storj off of the testify lib
entirely, but at least we should fix these instances. While doing
testing for the openssl port, this particular set of tests caused me
very much trouble, because a thing would fail, and then the test would
hang instead of telling me what the error was. And then once i figured
out the actual error and fixed it, the next thing would fail and I'd
have to do it all over again.