Fix CalcEncryptedSize and CalcPieceSize
When encrypting and erasure coding data, the data is padded to be
divisible by the encoded block size or stripe size, respectively. This
padding adds 4 bytes at then end of the padded data with the number of
bytes that have been added (the padding bytes).

These 4 bytes were not taken into account byte the CalcEncryptedSize and
CalcPieceSize function, so they produced wrong result in some corner
cases.