test-network-stalls tests... network stalls
in particular, right now, it just tests whether an uplink correctly
times out after some amount of time when one of the nodes it's talking
to suddenly goes offline.

This tool is meant to be run under `storj-sim network test`.

I am _very_ open to suggestions as to the following:

 * where should this live? I've put it in scripts/ right now, but I'm
   not sure it fits there, as I don't know exactly what we intended
   scripts/ to be. Tools? Test scripts? Both?

 * should this actually be written to use the golang 'testing' module?
   It kind of seems like it should, because it is a test after all, but
   I don't know how that would work.

 * should we add this to scripts/test-sim.sh?