add items to ObjectMeta, UploadOptions, BucketConfig
It seems to make sense for each of these to have a RedundancyScheme, a
DataCipher, and an EncryptionBlockSize. We could use
storj.EncryptionScheme for the latter two, but we want a Cipher type
where we can distinguish an unset value (0) from storj.Unencrypted (0).
We could just use a *storj.EncryptionScheme and have the nil pointer
mean to use defaults, but we would still have the same problem with
PathCipher. It doesn't make sense for PathCipher to use
storj.EncryptionScheme because it doesn't have a BlockSize. We could use
uplink.Cipher for PathCipher and storj.EncryptionScheme for DataCipher,
but the mismatch is mentally jarring.