use SerializableMeta to store bucket attributes
It doesn't work well to use the bucket-level pointer to store the
default redundancy strategy, as such pointers are inline pointers and
don't have a RemoteSegment field at all. Also, it was proving to be an
enormous pain to plumb the information both up and down through existing
APIs without changing too much else.

So instead, this change undoes most of commit e4a70e3, and makes use of
the arbitrary-string-key-value map in the SerializableMeta structure.

This strategy has the obvious drawback of requiring a bunch of integer
string formatting and string parsing every time we check the metadata
on a bucket (which is, I think, every time we get any object). If it
proves to be a performance problem, we can revisit.