cmd/uplink: Ref default config filename
The cmd/uplink binary use the default filename that the pkg/process uses
to read/write the configuration parameters, however it doesn't reference
to it because pkg/process don't export it.

This commit export the pkg/process of the configuration default filename
and change the cmd/uplink for referencing it rather than hard coding it.

These changes will avoid that if such filename change in the pkg/process
the cmd/uplink won't stop to work because of it.