cmd/uplink: save root enc key into separated file
Save the root encryption key used for encrypting the data (a.k.a
passphrase) into a separated file instead of saving it into the
configuration file.

For avoiding to break the current API of uplink package, the uplink
configuration file saves the file path to the file which contains the
encryption key.