don't encrypt bucket info
Background: see https://github.com/storj/storj/pull/1761

Summary: we should not require an encryption key to manage buckets
(though we require an encryption key to manage objects within a
bucket), so that bucket management can be done in the web interface.
This means the encryption key used when opening a project is going away.

We have a network wipe planned for next release, so for that wipe let's
make sure data and buckets uploaded after that wipe do not require an
encryption key.

Change-Id: If5ddd4e6bf90e38c7aaa55b518e1d62af71704b9