cmd/gateway: use encryption key file
Because cmd/uplink configuration file has been changed for not
containing the encryption key value but a path to a file which holds the
key, the gateway should follow the same rule.

This commit change the gateway configuration file as it has been changed
for cmd/uplink and remove the possibility to run the gateway passing the
encryption key by flag.

Because of the changes in the cmd/gateway, the cmd/storj-sim has had to
be adapted.