uplink: remove unneeded err value
An err value was added because the attempt to run the cmd/uplink without
encryption key, however later on the key has been determined as required
because the cmd/uplink was updated in upstream for using the lib/uplink
hence the need of identifying such error has disappeared.

This commit remove such error value and leave the places where it was
returned using the common way of creating an error.