uplink: remove Key field from EncryptionConfig type
Remove the Key field from the EncryptionConfig type because the key has
been moved to be stored in a file.

Because of that, cmd/uplink, cmd/gateway and an integration test of
pkg/miniogw needed to be changed to load the key from the file rather
than accessing it through the same EncryptionConfig type.