pkg/audit: Readdress ignored verifier tests
The Verifier type of the pkg/audit had a test which was totally ignored
due to a +build tag.

Ignoring test using a +build is confusing because it isn't clear that
the test is fully ignored and the test doesn't get compiled anymore so
all the breaking changes in the implementation don't warm to update the
test.

This changes remove the +build tag and resuscitate the tests which
haven't required that much changes, the rest of them have been deleted.