add capability of logging all GRPC calls/payloads
If the server config has DebugLogTraffic set, info about every grpc call
will be logged. these logs include peer address, which service was
contacted, the method name, the contents of the RPC payload, and whether
an error was encountered while handling the call.

The intent is to capture logs like this on a production satellite for an
interval of time, and then use that information to replay traffic
against testing networks under controlled conditions to better
investigate performance issues.

When DebugLogTraffic is not set, there does not appear to be a
discernible impact on performance compared to that seen before this
commit.

When DebugLogTraffic _is_ set, there is an impact on performance, of
course. scripts/test-sim.sh takes about 9% more time to complete, and
the satellite is only a portion of the work done there, so the real
overhead on a satellite by itself is likely higher than that. I'm not
sure how to benchmark it effectively, though, without the data I'm
trying to make available here.