lib/uplink: remove NewReader
there's a few reasons for this but the main one is:
a seekable interface simply doesn't map well to the semantics
of loading large streams of data over the network.
good interfaces match the semantics of what the computer is doing,
and a seekable interface has a surprising amount of subtle
challenges when trying to optimize performance. as lib/uplink is
our public interface, i'm not eager to give our downstream users
any ideas or try to do anything this early that might lead anyone
to believe we will efficiently support randomized seeking.

this PR takes a chapter out of gocloud.dev/blob's interface and
simply does not provide a seekable type. Instead, where
gocloud.dev/blob offers a NewRangeReader, we stick with the
similar DownloadRange method.

Change-Id: I65fe739411068a576d8ad787f0f7688c4d649ea3