Update password based key derivation design doc
This brings it in line with what is actually implemented, giving
up on bucket renames, etc. Renames can be implemented by having a
"shadow name" on top of the actual bucket name, at the cost of
an indirection.

This also changes the design slightly to only export one function
to derive keys, since we fold the bucket name into any key
derivations that start at the empty path, removing the need to have
a separate argon2 step to derive a key for a bucket from a default
passphrase.

Change-Id: Iaa0491cc97a44d8a465f10051f4d8e6eb1104c43