overlay cache: slight modification of node-is-online rules
the proposition is that we aren't always consistent at pinging
nodes in a regular fashion. either the satellite was not running,
or the discovery refresh was not operating at the regular schedule.
we will consider a node online if it was successfully contacted in
the last hour, even if it has failed most recently (and perhaps
temporarily). we will also consider a node offline if our most
recent interaction was successful, regardless of if the most
recent interaction was recent.

Change-Id: I493a1a7c92a7f2e5c0f1344f7e888d32209838ea