add OrderLimit to PieceHeader
and add FormatVersion to PieceHeader too

also get rid of the pieces.PieceHeader struct; it is no longer useful
for what I originally wanted. we can pass pb.PieceHeader around without
any problems in the places where it is needed directly.

rename Writer.size -> Writer.dataSize

store PieceHash.Timestamp and not OrderLimit.OrderCreation for
PieceHeader.CreationTime, because I don't see that we require those to
be the same value, and the thing we care about here is explicitly
the PieceHash.Timestamp from the uplink.