don't report total recovered bytes after Collect
it is too expensive or complicated to get an accurate count of bytes
recovered when we don't have sizes stored in the pieceinfo database.
we could stat the piece files right before deleting them, but then
there are a whole lot of cases we need to handle because the stat and
delete will not be done atomically together.

reporting the number of collected pieces in the log ought to be
sufficient. if it's not, maybe we could report a 'bytes available'
from the volume before and after, as an approximation.