satellite/auth: add the concept of an admin key
this change makes administrator keys/macaroons that are off by
default but can be specified in satellite config. If a satellite
is configured with an admin macaroon head and secret, macaroons
constructed with that head and secret will be allowed to specify
their project id and work with any project.

Change-Id: I0dcbab881b2797f334e0d543827bbc8b6c7fcde6