piecestore: update usage before completing upload
The upload code currently updates the usage in a deferred call to saveOrder().
The consequence is that in the success case, the RPC is completed before
the usage has been updated.

This change repurposes the deferred call to update usage in the
failure case, while explicitly updating the usage before completing the
RPC.

This fixes some test flakiness when using dRPC
(TestOrderLimitGetValidation) since dRPC was quicker to issue a
follow-up download RPC that was expected to fail due to low bandwidth
but was succeeding because the deferred usage update had not taken
place. gRPC latency was such that the defer was always being executed in
time.